fix(l1_consistency_checker): decouple block intake from verification
The L1 consistency checker ran block intake and the CPU-heavy L1 commit
verification in a single pipeline task driven by one `select!`. During a
fresh EN sync, once the L1 watcher flushed a burst of historical commit
events, the synchronous commitment rebuild (~67ms/batch) saturated the
task and starved block intake. The terminal component fell more than one
channel-capacity (4096) behind TreeManager, so the pipeline's fail-fast
`try_send` panicked with "consumer is catastrophically behind" — even
though the cache was nowhere near its configured bound.

Split the responsibilities:

- `LocalBatchDataCacher`: a slim terminal pipeline component that only
  folds replayed blocks into the shared `TreeBlockCache`. It keeps the
  soft-bound backpressure (waits for capacity when full), which is now
  viable because eviction happens concurrently.
- `L1ConsistencyChecker`: a standalone task that consumes committed
  batches, waits for their blocks via the existing `wait_for_range`,
  rebuilds and compares the commitment, then evicts the verified prefix
  and advances the latest-verified watermark.

Both share the cache `watch::Sender`, so verification can no longer block
intake.
