+++ lib/l1_consistency_checker/src/lib.rs
@@ -1,5 +1,7 @@
 mod cache;
+mod cacher;
 mod checker;
 
 pub use cache::{TreeBlockCache, TreeBlockCacheReceiverExt};
+pub use cacher::LocalBatchDataCacher;
 pub use checker::{L1CommittedBatch, L1ConsistencyChecker};
